Modify Create,Update,Delete UseCases

+++ app/Application/Department/UseCases/DeleteDepartmentUseCase.php
@@ -4,6 +4,7 @@ namespace App\Application\Department\UseCases;
 
 use App\Domain\Department\Repositories\DepartmentRepositoryInterface;
 use App\Domain\Department\Entities\Department;
+use App\Http\Responses\ApiResponse;
 
 class DeleteDepartmentUseCase
 {
@@ -13,6 +14,12 @@ class DeleteDepartmentUseCase
 
     public function execute(int $id): void
     {
+        $department = $this->departmentRepository->findById($id);
+
+        if (!$department) {
+            throw new \DomainException("Department with ID [$id] not found.");
+        }
+
         $this->departmentRepository->delete($id);
     }
 }
